requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1
